+++ R/ard_survival_survfit.R
@@ -26,9 +26,9 @@
 #' @param y (`Surv` or `string`)\cr
 #'   an object of class `Surv` created using [survival::Surv()]. This object will be passed as the left-hand side of
 #'   the formula constructed and passed to [survival::survfit()]. This object can also be passed as a string.
-#' @param variables (`character`)\cr
+#' @param variables ([`tidy-select`][dplyr::dplyr_tidy_select])\cr
 #'   stratification variables to be passed as the right-hand side of the formula constructed and passed to
-#'   [survival::survfit()].
+#'   [survival::survfit()]. Default is `NULL` for an unstratified model, e.g. `Surv() ~ 1`.
 #' @param method.args (named `list`)\cr
 #'   named list of arguments that will be passed to [survival::survfit()].
 #' @inheritParams rlang::args_dots_empty
@@ -164,15 +164,15 @@ ard_survival_survfit.survfit <- function(x, times = NULL, probs = NULL, type = N
 
 #' @rdname ard_survival_survfit
 #' @export
-ard_survival_survfit.data.frame <- function(x, y, variables,
+ard_survival_survfit.data.frame <- function(x, y,
+                                            variables = NULL,
                                             times = NULL, probs = NULL, type = NULL,
                                             method.args = list(conf.int = 0.95), ...) {
   set_cli_abort_call()
 
   # check/process inputs -------------------------------------------------------
   check_not_missing(y)
-  check_not_missing(variables)
-  check_class(variables, "character")
+  cards::process_selectors(x, variables = {{ variables }})
 
   # process outcome as string --------------------------------------------------
   y <- enquo(y)
@@ -180,11 +180,25 @@ ard_survival_survfit.data.frame <- function(x, y, variables,
   if (tryCatch(is.character(eval_tidy(y)), error = \(e) FALSE)) y <- eval_tidy(y) # styler: off
   # otherwise, convert expr to string
   else y <- expr_deparse(quo_get_expr(y))  # styler: off
+  check_class(
+    with(x, eval(parse_expr(y))),
+    cls = "Surv",
+    message =
+      "The {.arg y} argument must be a string or expression that evaluates to an object of class {.cls Surv}
+     most often created with {.fun survival::Surv} or {.fun ggsurvfit::Surv_CNSR}."
+  )
+
 
   # build model ----------------------------------------------------------------
+  survfit_formula <-
+    case_switch(
+      !is_empty(variables) ~ stats::reformulate(termlabels = bt(variables), response = y),
+      .default = stats::reformulate(termlabels = "1", response = y)
+    )
+
   construct_model(
     data = x,
-    formula = stats::reformulate(termlabels = bt(variables), response = y),
+    formula = survfit_formula,
     method = "survfit",
     package = "survival",
     method.args = {{ method.args }}